Auto stash before rebase of "best-branch"
