+++ src/index.css
@@ -3,3 +3,7 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+a:hover {
+  cursor: pointer;
+}
